feat: standardize layout column styling, add new tailwind spacing helpers
