--- components/game-container/left-container.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function LeftColumn() {
-	return <div className="hidden xl:block absolute left-full top-0 ml-4 w-64">{/* Placeholder */}</div>;
-}
