+++ components/game-container/right-column.tsx
@@ -1,7 +1,7 @@
 import StatsPanel from '@/components/game-container/stats-panel';
 export default function RightColumn() {
 	return (
-		<div className="hidden xl:block absolute left-full top-0 ml-4 w-64">
+		<div className="hidden xl:block absolute left-full ml-layout-spacing top-layout-top w-64">
 			<StatsPanel />
 		</div>
 	);
